readme update

+++ src/main/java/org/laarryy/clippyv2/commands/moderation/PruneCommand.java
@@ -14,7 +14,7 @@ public class PruneCommand implements CommandExecutor {
 
     @Command(aliases = {"!prune"}, usage = "!prune <amount>", description = "Prunes a certain amount of messages (between 2 and 100)")
     public void onCommand(TextChannel channel, String[] args, MessageAuthor author, Message message) {
-        if (author.canKickUsersFromServer() && StringUtils.isNumeric(args[0])) {
+        if (author.canKickUsersFromServer() && StringUtils.isNumeric(args[0]) && !author.isYourself()) {
             int amount = Integer.parseInt(args[0])+1;
 
             channel.getMessages(amount).thenCompose(MessageSet::deleteAll).exceptionally(ExceptionLogger.get());
